refactor: model 추가 및 코드 정리
