+++ app/src/main/java/nextstep/shoppingcart/data/model/CartItem.kt
@@ -1,0 +1,8 @@
+package nextstep.shoppingcart.data.model
+
+data class CartItem(
+    val product: Product,
+    val count: Int,
+) {
+    val totalPrice: Long get() = product.price * count
+}
