+++ app/src/main/java/nextstep/shoppingcart/data/model/Product.kt
@@ -1,0 +1,11 @@
+package nextstep.shoppingcart.data.model
+
+import android.os.Parcelable
+import kotlinx.parcelize.Parcelize
+
+@Parcelize
+data class Product(
+    val name: String,
+    val imageUrl: String,
+    val price: Long,
+) : Parcelable
