+++ app/src/main/java/nextstep/shoppingcart/ui/component/ShoppingCartItem.kt
@@ -30,8 +30,8 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import coil.compose.AsyncImage
-import nextstep.shoppingcart.data.CartItem
-import nextstep.shoppingcart.data.Product
+import nextstep.shoppingcart.data.model.CartItem
+import nextstep.shoppingcart.data.model.Product
 import nextstep.shoppingcart.ui.theme.Grey10
 
 @Composable
