+++ app/src/main/java/nextstep/shoppingcart/ui/product/detail/ProductDetailContent.kt
@@ -20,7 +20,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import coil.compose.AsyncImage
-import nextstep.shoppingcart.data.Product
+import nextstep.shoppingcart.data.model.Product
 import nextstep.shoppingcart.ui.component.BlueBottomButton
 import nextstep.shoppingcart.ui.component.PriceLabel
 import nextstep.shoppingcart.ui.theme.BlackContent
@@ -29,7 +29,7 @@ import nextstep.shoppingcart.ui.theme.LineColor
 @Composable
 fun ProductDetailContent(
     product: Product,
-    onClickShoppingCartAddButton: () -> Unit,
+    onClickCartAddButton: () -> Unit,
     modifier: Modifier = Modifier,
 ) {
     Column(
@@ -60,7 +60,7 @@ fun ProductDetailContent(
 
         BlueBottomButton(
             label = "장바구니 담기",
-            onClick = onClickShoppingCartAddButton,
+            onClick = onClickCartAddButton,
             modifier = Modifier.fillMaxWidth()
         )
     }
